Conflicts:
	grails-app/conf/DataSource.groovy
	grails-app/controllers/species/groups/UserGroupController.groovy
	grails-app/domain/species/Resource.groovy
	grails-app/services/species/participation/DigestService.groovy
	grails-app/services/speciespage/ObservationService.groovy
	grails-app/views/emailtemplates/grantedPermission.gsp
	grails-app/views/index.gsp
	grails-app/views/species/_speciesFieldImageUpload.gsp
	web-app/js/species/species.js

+++ grails-app/controllers/species/groups/UserGroupController.groovy
@@ -612,8 +612,8 @@ class UserGroupController {
 			}
 			
 			String usernameFieldName = SpringSecurityUtils.securityConfig.userLookup.usernamePropertyName
-			def founders = userGroupInstance.getFounders(userGroupInstance.getFoundersCount(), 0L);
-            founders.addAll(userGroupInstance.getExperts(userGroupInstance.getExpertsCount(), 0L));
+		def founders = userGroupInstance.getFounders(userGroupInstance.getFoundersCount() as int, 0L);
+            founders.addAll(userGroupInstance.getExperts(userGroupInstance.getExpertsCount() as int, 0L));
             msg = messageSource.getMessage("default.confirm.membership", null, request.locale)
 			founders.each { founder ->
 				log.debug "Sending email to  founder ${founder}"
